requirements.txt
requests==2.31.0
telethon==1.34.0
python-dotenv==1.0.1
pandas_market_calendars==4.4.0
tenacity==8.2.3